[#2]docs: write README for step2 first trial
