+++ OpenMarket/OpenMarket/APINetwork/NetworkManager.swift
@@ -1,0 +1,8 @@
+//
+//  NetworkManager.swift
+//  OpenMarket
+//
+//  Created by Seungjin Baek on 2021/06/02.
+//
+
+import Foundation
